Create Server.py
SERVER SCRIPT


Server program has all the logic to control and regulate the Chat, so most of the chat logic is implemented with a server program. So first step of communication is to identify the users, how server do this? In network communication, users are identified by a socket which is nothing but a combination of IP address and port address. So, for human understanding, Alice and Bob will be chatting but for a network, it is two sockets process which is sending and receiving bytes.  Steps involved in this process is as follows:

*Create socket
*Communicate the socket address
*Keep waiting for an incoming connection request/s
*Connect to client
*Receive the message
*Decode the destination user and select the socket
*Send a message to the intended client
*Keep repeating step 5 & 6 as per users wish
*Exit i.e. end the communication by terminating the connection