client.py
Client script

Client script is run by the user, so the same client code will be run by a different user but each will have a separate socket so they will have their unique communication channel. 
Client script uses to be thin because it has very less work i.e. it only connect with the server and send and receive messages. The steps involved in client script are:



*Create a unique client socket per instance/user
*Connect to the server with given socket address (IP and port)
*Send and receive messages
*Repeat step 3 as per configuration
*Close the connection